requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2
psutil>=5.9.0
sacrebleu>=2.4.0
codecarbon>=2.3.6

